requirements.txt
feedparser
python-telegram-bot==20.3
PyGithub
python-dotenv
